fixed unsigned compares
git-svn-id: svn+ssh://svn.lighttpd.net/lighttpd/branches/lighttpd-merge-1.4.x@855 152afb58-edef-0310-8abb-c4023f1b3aa9

+++ src/base.h
@@ -508,7 +508,7 @@ typedef struct server {
 	int want_fds;   /* waiting fds */
 	int sockets_disabled;
 	
-	int max_conns;
+	size_t max_conns;
 
 	/* buffers */
 	buffer *parse_full_path;
